fix: reorganize folder structure for services part
