+++ src/services/get/problemStatement/getProblemStatement.ts
@@ -1,0 +1,1 @@
+export function getProblemStatement(yearDay: string, destination: string) {}
